feat!(transformation): impl `Default` trait + add `Copy` trait
others remarkable:
* add `inverse` method
* change `scale` signature
* refactor methods and traits
* improve unit tests

BREAKING CHANGES
* with `Default::default` trait no more `new` methods
* `test_mul_transformation` is now ignored because it fails!
